completed ex 17 taxi

+++ exercises/ex_16_k-Means/include/Point_Rn.hpp
@@ -6,7 +6,7 @@ using std::vector;
 
 class Point_Rn {
 protected: 
-    double dimensions;
+    int dimensions;
     vector<double> coordinates;
 
 public:
